+++ sentry-android-okhttp/src/main/java/io/sentry/android/okhttp/SentryOkHttpInterceptor.kt
@@ -5,6 +5,7 @@ import io.sentry.HubAdapter
 import io.sentry.IHub
 import io.sentry.ISpan
 import io.sentry.SpanStatus
+import io.sentry.TracingOrigins
 import java.io.IOException
 import okhttp3.Interceptor
 import okhttp3.Request
@@ -32,11 +33,13 @@ class SentryOkHttpInterceptor(
         var code: Int? = null
         try {
             val requestBuilder = request.newBuilder()
-            span?.toSentryTrace()?.let {
-                requestBuilder.addHeader(it.name, it.value)
-            }
-            span?.toTraceStateHeader()?.let {
-                requestBuilder.addHeader(it.name, it.value)
+            if (span != null && TracingOrigins.contain(hub.options.tracingOrigins, request.url.toString())) {
+                span.toSentryTrace().let {
+                    requestBuilder.addHeader(it.name, it.value)
+                }
+                span.toTraceStateHeader()?.let {
+                    requestBuilder.addHeader(it.name, it.value)
+                }
             }
             request = requestBuilder.build()
             response = chain.proceed(request)